initial configuration slider
