+++ TaskAppsAssasinsIOS/Controller/Helper/AudioViewController.swift
@@ -10,21 +10,8 @@ import UIKit
 
 // TODO: Jay
 extension NoteDetailViewController {
-    
-    func loadRecordingUI() {
-        //recordTapped()
-        prepareForRecording()
 
-    }
-    
-    @objc func recordTapped() {
-        if audioRecorder == nil {
-            startRecording()
-        } else {
-            finishRecording(success: true)
-        }
-    }
- 
+    /*
     func stopRecording() {
         if let recorder = audioRecorder {
            if recorder.isRecording {
@@ -44,13 +31,8 @@ extension NoteDetailViewController {
            }
        }
     }
-    
-    
-    func saveRecordingToCoreData (title:String, comments:String) {
-
-    }
-    
-    //func startRecording() {
+    */
+    /*
     func prepareForRecording() {
         recordAudioButton.image = UIImage(systemName: "mic")
         
@@ -78,13 +60,15 @@ extension NoteDetailViewController {
             recordAudioButton.image = UIImage(systemName: "mic")
             finishRecording(success: false)
         }
-    }    
+    }
+    */
     
+    /*
     func getDocumentDirectory() -> URL {
         let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
         let documentsDirectory = paths[0]
         return documentsDirectory
-    }
+    }*/
     
     func finishRecording(success: Bool) {
         audioRecorder?.stop()
@@ -114,18 +98,101 @@ extension NoteDetailViewController {
     
     func audioRecorderEncodeErrorDidOccur(_ recorder: AVAudioRecorder, error: Error?) {
         // Show in UI, error happened in audio recording
-        //audioReocrding(with error: error)
         if error != nil {
+            let errorAlert = UIAlertController(title: "Recording Audio", message: error!.localizedDescription, preferredStyle: .alert)
+            errorAlert.addAction(UIAlertAction(title: "OK", style: .cancel))
+            present(errorAlert, animated: true)
             print("Recording audio error \(error!.localizedDescription)")
         }
     }
-     
       
     func finishRecording() {
         audioRecorder?.stop()
         audioRecorder = nil
     }
     
+    // MARK: ---------------------------------------------------------------------------------------------------------
+    
+    // TODO: Aswin
+    func startRecording() {
+        
+        let directoryURL = FileManager.default.urls(for: FileManager.SearchPathDirectory.documentDirectory, in:
+                    FileManager.SearchPathDomainMask.userDomainMask).first
+                
+        let audioFileName = UUID().uuidString + ".m4a"
+                let audioFileURL = directoryURL!.appendingPathComponent(audioFileName)
+                soundURL = audioFileName   // Sound URL to be stored in CoreData
+
+        let settings = [
+            AVFormatIDKey: Int(kAudioFormatMPEG4AAC),
+            AVSampleRateKey: 44100,
+            AVNumberOfChannelsKey: 2,
+            AVEncoderAudioQualityKey: AVAudioQuality.high.rawValue]
+
+        do {
+            audioRecorder = try AVAudioRecorder(url: audioFileURL, settings: settings)
+            audioRecorder?.delegate = self
+            audioRecorder?.record()
+            recordAudioButton.image = UIImage(systemName: "stop.circle.fill")
+        } catch {
+            finishRecording(success: false)
+        }
+    }
+    
+    func loadRecordingFuntionality() {
+
+        recordingSession = AVAudioSession.sharedInstance()
+
+        do {
+            try recordingSession.setCategory(.playAndRecord, mode: .default)
+            try recordingSession.setActive(true)
+            recordingSession.requestRecordPermission() { [unowned self] allowed in
+                DispatchQueue.main.async {
+                    if allowed {
+                        self.recordAudioButton.isEnabled = true
+                    } else {
+                        self.recordAudioButton.isEnabled = false
+                    }
+                }
+            }
+        } catch {
+            print("An error had ocurred configuring the Audio Rec functionality \(error.localizedDescription)")
+        }
+    }
+
+    // MARK: Play audio button
+    @objc func playAudio(_ sender: UIButton) {
+        if audioRecorder == nil {
+            do {
+                var documentPath = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true)[0]
+                
+                // pass the index of record to play
+                // get the absolute path and file name from the array (Database)
+                documentPath.append("/\(audioPath[sender.tag])")
+
+                let url = NSURL(fileURLWithPath: documentPath)
+                
+                try player = AVAudioPlayer(contentsOf: url as URL)
+                scrubber.maximumValue = Float(player!.duration)
+                player?.volume = 1.0
+                player?.play()
+                timer.invalidate()
+                print("Audio is playing \(String(describing: player?.isPlaying))")
+            } catch {
+                print(error.localizedDescription)
+                timer  = Timer.scheduledTimer(timeInterval: 1, target: self, selector: #selector(updateScrubber), userInfo: nil, repeats: true)
+            }
+        }
+    }
+
+    // Start recording audio
+    @objc func recordTapped() {
+        if audioRecorder == nil {
+            startRecording()
+        } else {
+            finishRecording(success: true)
+        }
+    }
 
 
 }
