+++ TaskAppsAssasinsIOS/View/AudioCustomTableViewCell.xib
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
     <dependencies>
-        <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -35,7 +33,7 @@
             <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
             <connections>
                 <outlet property="audioPlayButton" destination="zHu-VE-gRm" id="aBH-YD-ISk"/>
-                <outlet property="audioSlider" destination="YFV-iW-N2Q" id="WTK-rs-KeG"/>
+                <outlet property="scrubber" destination="YFV-iW-N2Q" id="WTK-rs-KeG"/>
             </connections>
             <point key="canvasLocation" x="112.97709923664122" y="19.718309859154932"/>
         </tableViewCell>
